fix typos for correct output

+++ 0x08-python-more_classes/2-rectangle.py
@@ -38,9 +38,9 @@ class Rectangle:
         self.__width = value
 
     def area(self):
-        return (self.__width * self.__height)
+        return int(self.__width) * int(self.__height)
 
     def perimeter(self):
         if self.__width == 0 or self.__height == 0:
             return 0
-        return ((self.__width * 2) + (self.__height * 2))
+        return int(self.__width * 2) + int(self.__height * 2)
